+++ src/main/java/org/truffleruby/core/regexp/ClassicRegexp.java
@@ -669,7 +669,8 @@ public class ClassicRegexp implements ReOptions {
     private static final int QUOTED_V = 11;
 
     @TruffleBoundary
-    public static Rope quote19(Rope bs, boolean asciiOnly) {
+    public static Rope quote19(Rope bs) {
+        final boolean asciiOnly = bs.isAsciiOnly();
         int p = 0;
         int end = bs.byteLength();
         final byte[] bytes = bs.getBytes();
@@ -721,7 +722,7 @@ public class ClassicRegexp implements ReOptions {
                 p += cl;
             }
             if (asciiOnly) {
-                return bs.withEncoding(USASCIIEncoding.INSTANCE, CR_7BIT);
+                return RopeOperations.withEncoding(bs, USASCIIEncoding.INSTANCE);
             }
             return bs;
         } while (false);