+++ src/main/java/org/truffleruby/core/regexp/RegexpNodes.java
@@ -211,8 +211,7 @@ public abstract class RegexpNodes {
         @Specialization(guards = "isRubyString(raw)")
         protected DynamicObject quoteString(DynamicObject raw) {
             final Rope rope = StringOperations.rope(raw);
-            boolean isAsciiOnly = rope.getEncoding().isAsciiCompatible() && rope.getCodeRange() == CodeRange.CR_7BIT;
-            return getMakeStringNode().fromRope(ClassicRegexp.quote19(rope, isAsciiOnly));
+            return getMakeStringNode().fromRope(ClassicRegexp.quote19(rope));
         }
 
         @Specialization(guards = "isRubySymbol(raw)")